update Gitignore
Going to keep the code that actually is basically a game hack hidden because I don't trust people to not think that they can take this code and use it for an actual hack. The game launches with -novac so I can be crude in how I access the games memory, so it would most definitely get you banned if you tried to copy it.  Hacking is not modding